fallback to -O3 if $(CC) != "gcc"
if CC isn't found on the enviornment variables, we use `cc` as a generic
compiler. `cc` is typically a symlink to gcc on most distros, but on
certain distros it can be symlink to clang instead.

since the default optimization flags used in the Makefile contains
various gcc specific flags, if CC wasn't explictly gcc all optimization
flags would get disabled so that the build doesn't fail (on clang and
other hipster compilers).

instead of disabling all optimization flags all together, now we
fallback to `-O3` instead. `-O` flag is specified by POSIX in c99 (1p)
so it should be portable across various compilers.

Fixes: https://github.com/N-R-K/sxcs/pull/3