rm preview images from the repo
commiting images into git was utterly retarded to begin with.

but in any case, this should make the "tagged" tarball smaller now.

people cloning the git repo are still doomed though... unless they clone
with something like `--depth ${n}`.

imgbox SEEMS to be around for a while and SEEMS to be reliable. but if
it turns out otherwise, then i think i can just create an orphan branch
and push it over to github to burn micro$oft's money.