change fprintf() -> fputs()
it always bothered me a lot how certain stdio functions take the
`FILE *` as the first arg (e.g fprintf) while others take it as the last
(e.g fputs, fgets). i believe that all functions which operate on file
stream should just take it as their first arg unanimously.

BUT even if they all took it as the last param, i wouldn't be too
bothered. except that's when you realize that functions like fprintf
CANNOT take it as the last param because it's a variadic function and
the va args have to be the last:

	fprintf(const char *fmt, ..., FILE *stream); // cant do this...

okay, perhaps that's the reason for this inconsistency. so now we got a
formula - if it's a va function the file stream goes first - otherwise
it goes last.

the situation isn't ideal, but it's good enough formula to make me not
bothered anymore. except there's one problem; the formula doesn't work.

lo and belhold:

	int fseek(FILE *stream, long offset, int whence);
	int fgetpos(FILE *restrict stream, fpos_t *restrict pos);
	void setbuf(FILE *restrict stream, char *restrict buf);

and so on...

perhaps there's some reason it was done this way. but to me, this just
looks like an absolute mess and i hate it.