WIP: Use XSyncAlarm instead of poll
Using `poll.3` worked in practice, but it's not a gurantee that has been
made on the interface contract (as pointed out by u/skeeto [0]); so it
makes sense to use XSyncAlarm instead.

This is a rough draft, but everything seems to be working fine.

One issue is that it doesn't deal with interrupts immediately as poll
does (`XNextEvent` will just go back to blocking). Thankfully the
default delay in sxcs is 16ms, so it's not _that_ big of a deal.

[0]: https://old.reddit.com/r/C_Programming/comments/xaswy8/software_renderer_not_working_efficiently_when/io4g1th/