make asserts opt-in rather than opt-out
since the time i've written sxcs, i've come to realize that having
assertions enabled by default is not a good idea. a lot of distros
and/or users do not add `-D NDEBUG` to their default set of compiler
flags.

a common convention is to use `-D DEBUG` to enable a debug build
instead. this way, release build is the default (as it should be) and
debug builds will require opt-in.