fix: black areas appearing on the right and bottom
For some zoom levels there is a black area on the right side and
the bottom of the magnifier. This is fixed by adding one pixel
in width and height to the magnified area.

You need `(float)MAG_SIZE / MAG_FACTOR` pixels to fill the area
of the magnifier glass. But `uint c` is this value rounded down
(by casting). Simply rounding up instead seems not to fix this
issue because this changes the offset as well.

The way the zoom currently works is by *pulling from* the source
image (rather than *pushing to* the dest) and so the extra pixel
in the source will just be ignored if needed.

Co-authored-by: NRK <nrk@disroot.org>
Closes: https://codeberg.org/NRK/sxcs/pulls/8